# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tack API",
    "description": "REST + WebSocket API for Tack, a lightweight, workflow-agnostic project-management tool. This contract is generated from the Rust handlers and domain models; it is the single source of truth for the wire format. All failing responses share the `{ \"error\": { \"status\", \"message\" } }` envelope.",
    "contact": {
      "name": "Tack",
      "email": "info@yielab.com"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0-beta.6"
  },
  "paths": {
...
    "/api/items/{item_id}/dependencies": {
      "get": {
        "tags": [
          "dependencies"
        ],
        "operationId": "list_dependencies",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "description": "Item ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dependency edges for the item",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Dependency"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "dependencies"
        ],
        "operationId": "create_dependency",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "description": "Source item ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDependency"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Dependency created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dependency"
                }
              }
            }
          },
          "400": {
            "description": "Cycle detected or duplicate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "operationId": "list_projects",
        "responses": {
          "200": {
            "description": "All projects in the workspace",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "operationId": "create_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/import": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "POST /api/projects/import",
        "description": "Accepts the same snapshot shape produced by `export?format=json|yaml`. The\nbody is parsed as YAML when the `Content-Type` mentions YAML, otherwise JSON —\nso a YAML export round-trips back in unchanged. (YAML is a JSON superset, so\nboth decode into the same intermediate `serde_json::Value`.)",
        "operationId": "import_project",
        "requestBody": {
          "description": "Project snapshot (JSON, or YAML when Content-Type mentions yaml) as produced by export",
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Import result with the new project and stats",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Invalid import payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "operationId": "get_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "operationId": "delete_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "projects"
        ],
        "operationId": "update_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProject"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated project",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{id}/export": {
      "get": {
        "tags": [
          "export"
        ],
        "summary": "GET /api/projects/:id/export",
        "operationId": "export_project",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Export file (JSON, YAML, or CSV per the `format` query)"
          },
          "400": {
            "description": "Unsupported format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{id}/import-csv": {
      "post": {
        "tags": [
          "import"
        ],
        "summary": "POST /api/projects/:id/import-csv",
        "description": "Accepts a UTF-8 CSV body (Content-Type: text/csv or text/plain).\nFirst row must be column headers.  Only `title` is required; all other\ncolumns are optional.  Recognised headers (case-insensitive):\n  title, description, type, item_type, status, priority, assignee, estimate\n\nReturns `{ \"created\": N, \"skipped\": N }`.",
        "operationId": "import_csv",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "CSV with a header row; `title` column required",
          "content": {
            "text/csv": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Counts of created and skipped rows",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "400": {
            "description": "Malformed CSV",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{project_id}/sprints": {
      "get": {
        "tags": [
          "sprints"
        ],
        "operationId": "list_sprints",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sprints for the project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sprint"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "sprints"
        ],
        "operationId": "create_sprint",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSprint"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Sprint created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sprint"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDependency": {
        "type": "object",
        "required": [
          "target_item_id",
          "dependency_type"
        ],
        "properties": {
          "dependency_type": {
            "$ref": "#/components/schemas/DependencyType"
          },
          "target_item_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "CreateProject": {
        "type": "object",
        "required": [
          "name",
          "project_type"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "project_type": {
            "$ref": "#/components/schemas/ProjectType"
          },
          "template": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Dependency": {
        "type": "object",
        "required": [
          "id",
          "source_item_id",
          "target_item_id",
          "dependency_type",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "dependency_type": {
            "$ref": "#/components/schemas/DependencyType"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "source_item_id": {
            "type": "string",
            "format": "uuid"
          },
          "target_item_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "DependencyType": {
        "type": "string",
        "enum": [
          "blocks",
          "is_blocked_by",
          "relates_to",
          "duplicates"
        ]
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "description": "Structured error envelope returned by every failing endpoint — see\n`crate::error::ApiError`. The body is always `{ \"error\": { status, message } }`.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "ProjectType": {
        "type": "string",
        "enum": [
          "software",
          "web",
          "mobile",
          "construction",
          "personal",
          "homework",
          "maintenance",
          "legal",
          "research",
          "event",
          "custom"
        ]
      },
...
    }
  },
...
}